+++ packages/sdk/CHANGELOG.md
@@ -23,6 +23,7 @@
 - `PendingUpdate.effect` now has the type `Upward<PendingUpateEffect>`.
 - `PassiveCommitteeInfo` now has been wrapped in `Upward`.
 - `NodeInfoConsensusStatus` and `NodeCatchupStatus` now have been wrapped in `Upward`.
+- `ElectionInfo` now has been wrapped in `Upward`
 
 #### `ConcordiumGRPCClient`: